TASK-17 Refactor/Create unit tests for localization file
